+++ lib/core/ui-basic/ui-basic-textarea/ui-basic-textarea.component.ts
@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { Component, HostBinding } from '@angular/core';
+import { Component } from '@angular/core';
 import { FormControl, ReactiveFormsModule } from '@angular/forms';
 import { CustomControlComponent } from '../../components/custom-control/custom-control.component';
 import {
@@ -27,9 +27,11 @@ import { providePropsBinding } from '../../providers/props-binding.provider';
   ],
   templateUrl: './ui-basic-textarea.component.html',
   styles: [],
+  host: {
+    class: 'ui-basic',
+  },
 })
 export class UiBasicTextareaComponent extends CustomControlComponent {
-  @HostBinding('class') hostClass = 'ui-basic';
   override control = new FormControl('');
 
   onChange?: any;